Time: 3 ms (29.22%), Space: 8.7 MB (20.91%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long int n=0;
+        long long int n=0;
         bool sign=false;
         if(x<0){bool sign=true;}
         while(x!=0){
